Вынес Datasource и Delegat в отдельные файлы.
